added update-king-index function
function is used to update king's index to specific place on the board. The
index is located outside game board so it won't interfere with the game.
This is done so we don't need to search for the king while we check if the
king is threatened. Should give some nice boost in performance.