refactored threaten-by-king?
This should be working version now.
First it sees if any of the surrounding board squares around index contain
opponents king.
If enemy king is found, then if our piece is king its under threat
automatically, otherwise 'move' enemy king to our square and test if it
instead is threatened.